requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djongo==1.3.6
idna==2.10
pkg-resources==0.0.0
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
urllib3==1.26.5
ipdb==0.13.9